+++ schwifty-ts/src/index.ts
@@ -2,6 +2,16 @@ export { BBAN } from "./bban.ts";
 export { BIC } from "./bic.ts";
 export { algorithms, getAlgorithm } from "./checksum/index.ts";
 export { clean, toAscii } from "./common.ts";
-export { Component } from "./domain.ts";
+export type { Bank, IBANSpec } from "./domain.ts";
+export { Component, Range } from "./domain.ts";
 export * from "./exceptions.ts";
-export { convertBbanSpecToRegex, IBAN } from "./iban.ts";
+export { IBAN } from "./iban.ts";
+export {
+  convertBbanSpecToRegex,
+  getAllBanks,
+  getBanksByBic,
+  getBanksByCode,
+  getBanksByCountry,
+  getCountries,
+  getIbanSpec,
+} from "./registry.ts";